refactor: fix extension debugging
