+++ packages/catppuccin-vsc/package.json
@@ -184,6 +184,7 @@
   "scripts": {
     "core:build": "tsx build.ts",
     "core:dev": "tsx build.ts --dev",
+    "core:watch": "tsx watch build.ts --dev",
     "schema:ui": "tsx src/hooks/updateSchemas.ts"
   },
   "homepage": "https://github.com/catppuccin/vscode/tree/main/packages/catppuccin-vsc#readme"
